requirements.txt
python-dotenv
uvicorn
langgraph
langchain[openai]
langchain-deepseek
langchain-mcp-adapters
langgraph-cli[inmem]